execution/bots/requirements.txt
# Project Apex - Python Dependencies
# Install with: pip install -r requirements.txt

# ============================================
# TELEGRAM BOT API (Core Communication)
# ============================================
python-telegram-bot==20.7
# Modern async Telegram bot framework
# Docs: https://docs.python-telegram-bot.org/

# ============================================
# AI MODEL ROUTING
# ============================================
litellm==1.30.0
# Universal LLM API - supports Ollama, Gemini, Claude, OpenAI
# Docs: https://docs.litellm.ai/

# ============================================
# SALESFORCE INTEGRATION
# ============================================
simple-salesforce==1.12.5
# Salesforce REST API client
# Note: sf CLI is separate (installed via npm)
# Docs: https://github.com/simple-salesforce/simple-salesforce

# ============================================
# NOTION INTEGRATION
# ============================================
notion-client==2.2.1
# Official Notion SDK for Python
# Docs: https://github.com/ramnes/notion-sdk-py

# ============================================
# DATABASE & CACHING
# ============================================
# SQLite is built-in to Python (no package needed)
aiosqlite==0.19.0
# Async SQLite for concurrent operations

redis==5.0.1
# Redis client for message queue
# Docs: https://redis-py.readthedocs.io/

# ============================================
# HTTP CLIENTS
# ============================================
httpx==0.26.0
# Modern async HTTP client (for webhooks, API calls)
# Docs: https://www.python-httpx.org/

aiohttp==3.9.1
# Alternative async HTTP client
# Docs: https://docs.aiohttp.org/

# ============================================
# SCHEDULING
# ============================================
APScheduler==3.10.4
# Advanced Python scheduler for background tasks
# Docs: https://apscheduler.readthedocs.io/

# ============================================
# WEB SCRAPING (Procurement Portal Scanner)
# ============================================
beautifulsoup4==4.12.2
# HTML parsing
# Docs: https://www.crummy.com/software/BeautifulSoup/

lxml==5.0.0
# Fast XML/HTML parser

# ============================================
# MONITORING & METRICS
# ============================================
prometheus-client==0.19.0
# Prometheus metrics exporter
# Docs: https://github.com/prometheus/client_python

psutil==5.9.6
# System resource monitoring (CPU, memory, disk)
# Docs: https://psutil.readthedocs.io/

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
# Load environment variables from .env
# Docs: https://saurabh-kumar.com/python-dotenv/

pydantic==2.5.3
# Data validation and settings management
# Docs: https://docs.pydantic.dev/

tenacity==8.2.3
# Retry logic with exponential backoff
# Docs: https://tenacity.readthedocs.io/

# ============================================
# DATE/TIME
# ============================================
python-dateutil==2.8.2
# Extended date/time utilities

pytz==2023.3
# Timezone support (Arizona doesn't observe DST!)

# ============================================
# LOGGING & DEBUGGING
# ============================================
structlog==24.1.0
# Structured logging
# Docs: https://www.structlog.org/

# ============================================
# TESTING (Optional - for development)
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# ============================================
# NOTES
# ============================================
#
# INSTALLATION:
#   pip install -r requirements.txt
#
# DOCKER:
#   These will be installed in Docker containers automatically
#
# VIRTUAL ENVIRONMENT (Recommended):
#   python3 -m venv apex_env
#   source apex_env/bin/activate
#   pip install -r requirements.txt
#
# SYSTEM REQUIREMENTS:
#   - Python 3.11 or higher
#   - sf CLI (Salesforce CLI) installed separately
#   - Ollama installed for local AI (optional)
#
# ============================================
